requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
lightgbm>=4.5
xgboost>=2.0
matplotlib>=3.8
plotly>=5.20
streamlit>=1.37
pyyaml>=6.0
shap>=0.45
joblib>=1.4
